Refactor pull to include ActionConfig #440
